+++ components/Footer/Footer.js
@@ -46,7 +46,7 @@ export default function Footer(props) {
               target="_blank"
               className={classes.a}
             >
-              Creative Tim
+              Onus
             </a>
             , made with love for a better web
           </span>